pyproject.toml
[project]
name = "customer-curn-prediction-asistant-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "joblib>=1.5.3",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "loguru>=0.7.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic-settings>=2.13.1",
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
]
